intellidraft/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-docx
python-multipart
requests
